Acts more like an NES
Now the parameter CPU_RESET_ADDRESS points to a location where the
address of the first instruction is. The CPU automatically loads in the
first two bytes at the location of the macro and then jumps there. It
may or may not be a hack where I just trick the CPU into executing an
actual jump command, but a processor is just a rock we trick into
thinking, so don't judge me too hard.